requirements.txt
pytest==2.9.1
hexdump==3.3
PyQt5==5.8.1
PyHamcrest